mysite/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-filter==23.4
djangorestframework==3.14.0
dnspython==2.4.2
idna==3.6
importlib-metadata==6.8.0
kafka-python==2.0.2
Markdown==3.5.1
pymongo==4.6.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sentry-sdk==1.36.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
zipp==3.17.0
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-filter==23.4
djangorestframework==3.14.0
dnspython==2.4.2
gunicorn==21.2.0
idna==3.6
importlib-metadata==6.8.0
kafka-python==2.0.2
Markdown==3.5.1
packaging==23.2
pymongo==4.6.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sentry-sdk==1.36.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
zipp==3.17.0
